fix: update evaluation utils and ci scripts
