+++ graph_datasets/utils/evaluation/eval_tools.py
@@ -1,32 +1,29 @@
-from typing import Dict
-
-import random
+"""Utils for evaluation.
+"""
+# pylint: disable=invalid-name,invalid-name,too-many-locals
 import os
-import scipy.sparse as sp
-import networkx as nx
+import random
+
 import numpy as np
-import pandas as pd
-from sklearn import preprocessing
-
-from munkres import Munkres
-from sklearn.metrics import accuracy_score, f1_score, normalized_mutual_info_score, adjusted_rand_score
-from sklearn.svm import LinearSVC
-from sklearn.cluster import KMeans
-from sklearn import linear_model as sk_lm
-from sklearn import metrics as sk_mtr
-from sklearn import model_selection as sk_ms
-from sklearn import multiclass as sk_mc
-from sklearn import preprocessing as sk_prep
-
 import torch
-from torch_geometric.data import Data
+from munkres import Munkres
 from six.moves import cPickle as pickle
+from sklearn.cluster import KMeans
+from sklearn.metrics import accuracy_score as ACC
+from sklearn.metrics import adjusted_mutual_info_score as AMI
+from sklearn.metrics import adjusted_rand_score as ARI
+from sklearn.metrics import f1_score as F1
+from sklearn.metrics import normalized_mutual_info_score as NMI
+from sklearn.svm import LinearSVC
+
+from ..common import tab_printer
 
 
 def load_dict(filename_):
-    with open(filename_, 'rb') as f:
+    with open(filename_, "rb") as f:
         ret_di = pickle.load(f)
     return ret_di
+
 
 def save_dict(di_, filename_):
     # Get the directory path from the filename
@@ -36,43 +33,39 @@ def save_dict(di_, filename_):
     if not os.path.exists(dir_path):
         os.makedirs(dir_path)
 
-    with open(filename_, 'wb') as f:
+    with open(filename_, "wb") as f:
         pickle.dump(di_, f)
+
 
 def split_train_test_nodes(data, train_ratio, valid_ratio, data_name, split_id=0, fixed_split=True):
     if fixed_split:
-        file_path = f'../input/fixed_splits/{data_name}-{train_ratio}-{valid_ratio}-splits.npy'
+        file_path = f"../input/fixed_splits/{data_name}-{train_ratio}-{valid_ratio}-splits.npy"
         if not os.path.exists(file_path):
-            print('There is no generated fixed splits')
-            print('Generating fixed splits...')
+            print("There is no generated fixed splits")
+            print("Generating fixed splits...")
             splits = {}
             for idx in range(10):
                 # set up train val and test
                 shuffle = list(range(data.num_nodes))
                 random.shuffle(shuffle)
-                train_nodes = shuffle[: int(data.num_nodes * train_ratio / 100)]
-                val_nodes = shuffle[int(data.num_nodes * train_ratio / 100):int(
-                    data.num_nodes * (train_ratio + valid_ratio) / 100)]
-                test_nodes = shuffle[int(data.num_nodes * (train_ratio+valid_ratio) / 100):]
-                splits[idx] = {
-                    'train': train_nodes,
-                    'valid': val_nodes,
-                    'test': test_nodes
-                }
-            save_dict(
-                di_=splits, filename_=file_path
-            )
+                train_nodes = shuffle[:int(data.num_nodes * train_ratio / 100)]
+                val_nodes = shuffle[
+                    int(data.num_nodes * train_ratio /
+                        100):int(data.num_nodes * (train_ratio + valid_ratio) / 100)]
+                test_nodes = shuffle[int(data.num_nodes * (train_ratio + valid_ratio) / 100):]
+                splits[idx] = {"train": train_nodes, "valid": val_nodes, "test": test_nodes}
+            save_dict(di_=splits, filename_=file_path)
         else:
             splits = load_dict(filename_=file_path)
         split = splits[split_id]
-        train_nodes, val_nodes, test_nodes = split['train'], split['valid'], split['test']
+        train_nodes, val_nodes, test_nodes = split["train"], split["valid"], split["test"]
     else:
         # set up train val and test
         shuffle = list(range(data.num_nodes))
         random.shuffle(shuffle)
-        train_nodes = shuffle[: int(data.num_nodes * train_ratio / 100)]
-        val_nodes = shuffle[
-                    int(data.num_nodes * train_ratio / 100):int(data.num_nodes * (train_ratio + valid_ratio) / 100)]
+        train_nodes = shuffle[:int(data.num_nodes * train_ratio / 100)]
+        val_nodes = shuffle[int(data.num_nodes * train_ratio /
+                                100):int(data.num_nodes * (train_ratio + valid_ratio) / 100)]
         test_nodes = shuffle[int(data.num_nodes * (train_ratio + valid_ratio) / 100):]
 
     return np.array(train_nodes), np.array(val_nodes), np.array(test_nodes)
@@ -80,26 +73,27 @@ def split_train_test_nodes(data, train_ratio, valid_ratio, data_name, split_id=0
 
 def cluster_eval(y_true, y_pred):
     """code source: https://github.com/bdy9527/SDCN"""
-    y_true = y_true.detach().cpu().numpy() if type(y_true) is torch.Tensor else y_true
-    y_pred = y_pred.detach().cpu().numpy() if type(y_pred) is torch.Tensor else y_pred
+    y_true = y_true.detach().cpu().numpy() if isinstance(y_true, torch.Tensor) else y_true
+    y_pred = y_pred.detach().cpu().numpy() if isinstance(y_pred, torch.Tensor) else y_pred
 
     l1 = list(set(y_true))
     numclass1 = len(l1)
     l2 = list(set(y_pred))
     numclass2 = len(l2)
 
+    # NOTE: NOT force to assign a random node into a missing class
     # fill out missing classes
-    ind = 0
-    if numclass1 != numclass2:
-        for i in l1:
-            if i in l2:
-                pass
-            else:
-                y_pred[ind] = i
-                ind += 1
+    # ind = 0
+    # if numclass1 != numclass2:
+    #     for i in l1:
+    #         if i in l2:
+    #             pass
+    #         else:
+    #             y_pred[ind] = i
+    #             ind += 1
 
-    l2 = list(set(y_pred))
-    numclass2 = len(l2)
+    # l2 = list(set(y_pred))
+    # numclass2 = len(l2)
 
     cost = np.zeros((numclass1, numclass2), dtype=int)
     for i, c1 in enumerate(l1):
@@ -123,45 +117,61 @@ def cluster_eval(y_true, y_pred):
         ai = [ind for ind, elm in enumerate(y_pred) if elm == c2]
         new_predict[ai] = c
 
-    acc = accuracy_score(y_true, new_predict)
-    f1_macro = f1_score(y_true, new_predict, average='macro')
+    acc = ACC(y_true, new_predict)
+    f1_macro = F1(y_true, new_predict, average="macro")
     return acc, f1_macro
 
 
-def unsup_eval(y_true, y_pred, epoch=0, quiet=False):
-    y_true = y_true.detach().cpu().numpy() if type(y_true) is torch.Tensor else y_true
-    y_pred = y_pred.detach().cpu().numpy() if type(y_pred) is torch.Tensor else y_pred
+def unsup_eval(y_true, y_pred):
+    y_true = y_true.detach().cpu().numpy() if isinstance(y_true, torch.Tensor) else y_true
+    y_pred = y_pred.detach().cpu().numpy() if isinstance(y_pred, torch.Tensor) else y_pred
 
-    acc, f1 = cluster_eval(y_true, y_pred)
-    nmi = normalized_mutual_info_score(y_true, y_pred, average_method='arithmetic')
-    ari = adjusted_rand_score(y_true, y_pred)
-    if not quiet:
-        print(epoch, ':acc {:.4f}'.format(acc), ', nmi {:.4f}'.format(nmi), ', ari {:.4f}'.format(ari),
-                ', f1 {:.4f}'.format(f1))
-    return acc, nmi, ari, f1
+    acc, macro_f1 = cluster_eval(y_true, y_pred)
+    nmi = NMI(y_true, y_pred, average_method="arithmetic")
+    ami = AMI(y_true, y_pred, average_method="arithmetic")
+    ari = ARI(y_true, y_pred)
+    return acc, nmi, ami, ari, macro_f1
 
 
-def kmeans_test(X, y, n_clusters, repeat=10, epoch=0, quiet=True):
-    y = y.detach().cpu().numpy() if type(y) is torch.Tensor else y
-    X = X.detach().cpu().numpy() if type(X) is torch.Tensor else X
+def kmeans_test(X, y, n_clusters, repeat=10, quiet=True):
+    y = y.detach().cpu().numpy() if isinstance(y, torch.Tensor) else y
+    X = X.detach().cpu().numpy() if isinstance(X, torch.Tensor) else X
 
     acc_list = []
     nmi_list = []
+    ami_list = []
     ari_list = []
     f1_list = []
     for _ in range(repeat):
         kmeans = KMeans(n_clusters=n_clusters)
         y_pred = kmeans.fit_predict(X)
-        acc_score, nmi_score, ari_score, macro_f1 = unsup_eval(
-            y_true=y, y_pred=y_pred,
-            epoch=epoch, quiet=quiet
+        (
+            acc_score,
+            nmi_score,
+            ami_score,
+            ari_score,
+            macro_f1,
+        ) = unsup_eval(
+            y_true=y,
+            y_pred=y_pred,
         )
         acc_list.append(acc_score)
         nmi_list.append(nmi_score)
+        ami_list.append(ami_score)
         ari_list.append(ari_score)
         f1_list.append(macro_f1)
-    return np.mean(acc_list), np.std(acc_list), np.mean(nmi_list), np.std(nmi_list), np.mean(ari_list), np.std(
-        ari_list), np.mean(f1_list), np.std(f1_list)
+    return (
+        np.mean(acc_list),
+        np.std(acc_list),
+        np.mean(nmi_list),
+        np.std(nmi_list),
+        np.mean(ami_list),
+        np.std(ami_list),
+        np.mean(ari_list),
+        np.std(ari_list),
+        np.mean(f1_list),
+        np.std(f1_list),
+    )
 
 
 def svm_test(data, embeddings, labels, train_ratios=(10, 20, 30, 40), repeat=10):
@@ -172,16 +182,19 @@ def svm_test(data, embeddings, labels, train_ratios=(10, 20, 30, 40), repeat=10)
         micro_f1_list = []
         for i in range(repeat):
             train_idx, val_idx, test_idx = split_train_test_nodes(
-                data=data, train_ratio=train_ratio, valid_ratio=train_ratio,
-                data_name=data.name, split_id=i
+                data=data,
+                train_ratio=train_ratio,
+                valid_ratio=train_ratio,
+                data_name=data.name,
+                split_id=i,
             )
             X_train, X_test = embeddings[np.concatenate([train_idx, val_idx])], embeddings[test_idx]
             y_train, y_test = labels[np.concatenate([train_idx, val_idx])], labels[test_idx]
             svm = LinearSVC(dual=False)
             svm.fit(X_train, y_train)
             y_pred = svm.predict(X_test)
-            macro_f1 = f1_score(y_test, y_pred, average='macro')
-            micro_f1 = f1_score(y_test, y_pred, average='micro')
+            macro_f1 = F1(y_test, y_pred, average="macro")
+            micro_f1 = F1(y_test, y_pred, average="micro")
             macro_f1_list.append(macro_f1)
             micro_f1_list.append(micro_f1)
         result_macro_f1_list.append((np.mean(macro_f1_list), np.std(macro_f1_list)))
@@ -190,8 +203,12 @@ def svm_test(data, embeddings, labels, train_ratios=(10, 20, 30, 40), repeat=10)
 
 
 def evaluate_results_nc(
-        data, embeddings, quiet=False, method='unsup',
-        alpha: float = 2.0, beta: float = 2.0,
+    data,
+    embeddings,
+    quiet=False,
+    method="unsup",
+    alpha: float = 2.0,
+    beta: float = 2.0,
 ):
     labels = data.y.detach().cpu().numpy()
     num_classes = data.num_classes
@@ -204,37 +221,85 @@ def evaluate_results_nc(
         else:
             embeddings = torch.cat((z_1, z_2), 1)
 
-    if method == 'both' or method == 'sup':
-        if not quiet:
-            print('SVM test')
-        svm_macro_f1_list, svm_micro_f1_list = svm_test(
-            data=data, embeddings=embeddings, labels=labels
+    if method in ("both", "sup"):
+        (
+            svm_macro_f1_list,
+            svm_micro_f1_list,
+        ) = svm_test(
+            data=data,
+            embeddings=embeddings,
+            labels=labels,
         )
         if not quiet:
-            print('Macro-F1: ' + ', '.join(
-                ['{:.2f}~{:.2f} ({:.1f})'.format(macro_f1_mean * 100, macro_f1_std * 100, train_size) for
-                 (macro_f1_mean, macro_f1_std), train_size in
-                 zip(svm_macro_f1_list, [10, 20, 30, 40])]))
-            print('Micro-F1: ' + ', '.join(
-                ['{:.2f}~{:.2f} ({:.1f})'.format(micro_f1_mean * 100, micro_f1_std * 100, train_size) for
-                 (micro_f1_mean, micro_f1_std), train_size in
-                 zip(svm_micro_f1_list, [10, 20, 30, 40])]))
+            print("SVM test")
+            tab_printer(
+                {
+                    "Macro F1":
+                        "\n".join(
+                            [
+                                f"{macro_f1_mean * 100:.2f}±{macro_f1_std * 100:.2f} ({train_size:.1f})"
+                                for (macro_f1_mean, macro_f1_std
+                                    ), train_size in zip(svm_macro_f1_list, [10, 20, 30, 40])
+                            ]
+                        ),
+                    "Micro F1":
+                        "\n".join(
+                            [
+                                f"{micro_f1_mean * 100:.2f}±{micro_f1_std * 100:.2f} ({train_size:.1f})"
+                                for (micro_f1_mean, micro_f1_std
+                                    ), train_size in zip(svm_micro_f1_list, [10, 20, 30, 40])
+                            ]
+                        ),
+                },
+                sort=False,
+            )
 
-    if method == 'both' or method == 'unsup':
-        if not quiet:
-            print('K-means test')
-        acc_mean, acc_std, nmi_mean, nmi_std, ari_mean, ari_std, f1_mean, f1_std = kmeans_test(
-            embeddings, labels, num_classes
+    if method in ("both", "unsup"):
+        (
+            acc_mean,
+            acc_std,
+            nmi_mean,
+            nmi_std,
+            ami_mean,
+            ami_std,
+            ari_mean,
+            ari_std,
+            f1_mean,
+            f1_std,
+        ) = kmeans_test(
+            embeddings,
+            labels,
+            num_classes,
         )
         if not quiet:
-            print('ACC: {:.2f}~{:.2f}'.format(acc_mean * 100, acc_std * 100))
-            print('NMI: {:.2f}~{:.2f}'.format(nmi_mean * 100, nmi_std * 100))
-            print('ARI: {:.2f}~{:.2f}'.format(ari_mean * 100, ari_std * 100))
-            print('F1: {:.2f}~{:.2f}'.format(f1_mean * 100, f1_std * 100))
+            print("K-means test")
+            tab_printer(
+                {
+                    "ACC": f"{acc_mean * 100:.2f}±{acc_std * 100:.2f}",
+                    "NMI": f"{nmi_mean * 100:.2f}±{nmi_std * 100:.2f}",
+                    "AMI": f"{ami_mean * 100:.2f}±{ami_std * 100:.2f}",
+                    "ARI": f"{ari_mean * 100:.2f}±{ari_std * 100:.2f}",
+                    "Macro F1": f"{f1_mean * 100:.2f}±{f1_std * 100:.2f}",
+                },
+                sort=False,
+            )
 
-    if method == 'sup':
+    if method == "sup":
         acc_mean = acc_std = nmi_mean = nmi_std = ari_mean = ari_std = f1_mean = f1_std = 0
-    elif method == 'unsup':
+    elif method == "unsup":
         svm_macro_f1_list = svm_micro_f1_list = [(0, 0), (0, 0), (0, 0), (0, 0)]
 
-    return svm_macro_f1_list, svm_micro_f1_list, acc_mean, acc_std, nmi_mean, nmi_std, ari_mean, ari_std, f1_mean, f1_std
+    return (
+        svm_macro_f1_list,
+        svm_micro_f1_list,
+        acc_mean,
+        acc_std,
+        nmi_mean,
+        nmi_std,
+        ami_mean,
+        ami_std,
+        ari_mean,
+        ari_std,
+        f1_mean,
+        f1_std,
+    )
